fix: run fist testgit add .!

+++ problem-3/src/Calculator.java
@@ -2,10 +2,11 @@
  * 
  * This class performs calculations
  */
+
 public class Calculator {
 
-    public long getMinDiv(long number) {
-        long div = 2;
+    public int getMinDiv(int number) {
+        int div = 2;
 
         for (div = 2; div * div < number; div++) {
             if (number % div == 0)
