+++ problem-3/src/CalculatorTest.java
@@ -5,7 +5,7 @@ public class CalculatorTest {
 	@Test
 	public void testAdd() {
 		Calculator calculator = new Calculator();
-		long result = calculator.getMinDiv(15);
+		int result = calculator.getMinDiv(15);
 		assertEquals(3, result, "15%3 == 0");
 	}
 }
